fixed problem with pf (single to uint8)

+++ @Particles/Particles.m
@@ -30,7 +30,7 @@ classdef Particles < handle
                 obj.nDiscreteStates = nDiscreteStates;
             else
                 obj.m = zeros(nParticles, nFrames, 'single');
-                obj.r = zeros(nParticles, nFrames, 'uint8');
+                obj.r = zeros(nParticles, nFrames, 'single');
             end
             obj.n = zeros(nParticles, nFrames, 'single');
             obj.weight = zeros(nParticles, 1);
